requirements.txt
matplotlib
numpy
Pillow
protobuf
scikit-learn
scipy
torchinfo
torchmetrics
torchvision
tqdm
wandb